refactor(frontend): divided code into different parts for better readability
